api/requirements.txt
flask-wtf==0.14.3
psycopg2-binary==2.9.1
werkzeug==0.16.1
markupsafe==2.0.1